backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
django-environ>=0.11
djangorestframework-simplejwt>=5.3